requirements.txt
# Dependências base (obrigatórias)
google-generativeai>=0.7.0
Pillow>=10.0.0
opencv-python>=4.9.0
numpy>=1.26.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# Dependências GPU local (opcional — instalar apenas se usar backend=local)
# pip install diffusers transformers accelerate
# pip install torch --index-url https://download.pytorch.org/whl/cu121

# API + Swagger (opcional)
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
